refactor(lsp): move client-state holders to lsp.client to break lsp↔providers cycle
