+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/lsp/client/ClientCapabilitiesHolder.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.lsp;
+package com.github._1c_syntax.bsl.languageserver.lsp.client;
 
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
